+++ tests/test_custom_checkers/test_invalid_range_index_checker.py
@@ -1,4 +1,5 @@
 import astroid
+import astroid.nodes as nodes
 import pylint.testutils
 
 from python_ta.checkers.invalid_range_index_checker import InvalidRangeIndexChecker
@@ -129,6 +130,14 @@ class TestInvalidRangeIndexChecker(pylint.testutils.CheckerTestCase):
         ):
             self.checker.visit_call(range_node)
 
+    def test_uninferable(self):
+        src = """
+        range(0, [][1])
+        """
+        range_node = astroid.extract_node(src)
+        with self.assertNoMessages():
+            self.checker.visit_call(range_node)
+
     def test_variables_undefined(self):
         src = """
         range(start, stop)  # These variables are undefined
@@ -141,17 +150,28 @@ class TestInvalidRangeIndexChecker(pylint.testutils.CheckerTestCase):
         src = """
         start = 1
         stop = 10
-        range(start, -stop)  # These variables can be inferred
+        range(start, -stop)
         """
         range_node = astroid.extract_node(src)
-        with self.assertAddsMessages(
-            pylint.testutils.MessageTest(
-                msg_id="invalid-range-index",
-                node=range_node,
-                args="4",
-            ),
-            ignore_position=True,
-        ):
+        # Even though these variables can have their values inferred,
+        # we are conservative and do not attempt inference (see test_variables_ambiguous below).
+        with self.assertNoMessages():
+            self.checker.visit_call(range_node)
+
+    def test_variables_ambiguous(self):
+        src = """
+        def f(numbers):
+            result = []
+            for _ in numbers:
+                result.append(1)
+
+            range(len(result))  #@
+        """
+        range_node = astroid.extract_node(src)
+        # In this case, result is currently being inferred as an empty list [],
+        # but may not be empty. So to be conservative we do not attempt to infer
+        # its result, and instead skip checking this range expression.
+        with self.assertNoMessages():
             self.checker.visit_call(range_node)